Move the classes in narrow_down.data_types were moved to narrow_down.storage.
This makes the library interface a little bit nicer.
