+++ narrow_down/__init__.py
@@ -4,4 +4,4 @@ __author__ = """Christian Krudewig"""
 __email__ = "chr1st1ank@krudewig-online.de"
 __version__ = "0.8.0"
 
-from . import data_types, hash, similarity_store, storage  # noqa
+from . import hash, similarity_store, storage  # noqa
